+++ src/module/features/customoracles.ts
@@ -1,60 +1,41 @@
-import { starforged, IOracle, IOracleCategory } from 'dataforged'
+import { starforged, ironsworn, IOracle, IOracleCategory } from 'dataforged'
 import { compact } from 'lodash'
 import { getFoundryISTableByDfId, getFoundrySFTableByDfId } from '../dataforged'
 import { cachedDocumentsForPack } from './pack-cache'
 
-export interface IOracleTreeNode {
+export interface OracleTreeNode {
   dataforgedNode?: IOracle | IOracleCategory
   tables: RollTable[]
   displayName: string
-  children: IOracleTreeNode[]
+  children: OracleTreeNode[]
   forceExpanded?: boolean
   forceHidden?: boolean
 }
 
 // For some reason, rollupJs mangles this
-const OracleCategories = starforged.default['Oracle Categories']
+const SFOracleCategories = starforged.default['Oracle Categories']
+const ISOracleCategories = ironsworn.default['Oracle Categories']
 
 const emptyNode = () =>
   ({
     displayName: '',
     tables: [],
     children: [],
-  } as IOracleTreeNode)
+  } as OracleTreeNode)
 
-export async function createIronswornOracleTree(): Promise<IOracleTreeNode> {
+async function createOracleTree(
+  compendium: string,
+  categories: IOracleCategory[],
+  tableGetter: (dfid: string) => Promise<StoredDocument<RollTable> | undefined>
+): Promise<OracleTreeNode> {
   const rootNode = emptyNode()
 
   // Make sure the compendium is loaded
-  await cachedDocumentsForPack('foundry-ironsworn.ironswornoracles')
+  await cachedDocumentsForPack(compendium)
 
   // Build the default tree
-  for (const category of OracleCategories) {
-    rootNode.children.push(
-      await walkOracleCategory(category, getFoundryISTableByDfId)
-    )
-  }
-
-  // Add in custom oracles from a well-known directory
-  await augmentWithFolderContents(rootNode)
-
-  // Fire the hook and allow extensions to modify the tree
-  await Hooks.call('ironswornOracles', rootNode)
-
-  return rootNode
-}
-
-export async function createStarforgedOracleTree(): Promise<IOracleTreeNode> {
-  const rootNode = emptyNode()
-
-  // Make sure the compendium is loaded
-  await cachedDocumentsForPack('foundry-ironsworn.starforgedoracles')
-
-  // Build the default tree
-  for (const category of OracleCategories) {
-    rootNode.children.push(
-      await walkOracleCategory(category, getFoundrySFTableByDfId)
-    )
+  for (const category of categories) {
+    rootNode.children.push(await walkOracleCategory(category, tableGetter))
   }
 
   // Add in custom oracles from a well-known directory
@@ -69,11 +50,27 @@ export async function createStarforgedOracleTree(): Promise<IOracleTreeNode> {
   return rootNode
 }
 
+export async function createIronswornOracleTree(): Promise<OracleTreeNode> {
+  return createOracleTree(
+    'foundry-ironsworn.ironswornoracles',
+    ISOracleCategories,
+    getFoundryISTableByDfId
+  )
+}
+
+export async function createStarforgedOracleTree(): Promise<OracleTreeNode> {
+  return createOracleTree(
+    'foundry-ironsworn.starforgedoracles',
+    SFOracleCategories,
+    getFoundrySFTableByDfId
+  )
+}
+
 async function walkOracleCategory(
   cat: IOracleCategory,
   tableGetter: typeof getFoundrySFTableByDfId
-): Promise<IOracleTreeNode> {
-  const node: IOracleTreeNode = {
+): Promise<OracleTreeNode> {
+  const node: OracleTreeNode = {
     ...emptyNode(),
     dataforgedNode: cat,
     displayName: game.i18n.localize(`IRONSWORN.OracleCategories.${cat.Name}`),
@@ -100,10 +97,10 @@ async function walkOracleCategory(
 async function walkOracle(
   oracle: IOracle,
   tableGetter: typeof getFoundrySFTableByDfId
-): Promise<IOracleTreeNode> {
+): Promise<OracleTreeNode> {
   const table = await tableGetter(oracle.$id)
 
-  const node: IOracleTreeNode = {
+  const node: OracleTreeNode = {
     ...emptyNode(),
     dataforgedNode: oracle,
     tables: compact([table]),
@@ -143,16 +140,16 @@ async function walkOracle(
   return node
 }
 
-async function augmentWithFolderContents(node: IOracleTreeNode) {
+async function augmentWithFolderContents(node: OracleTreeNode) {
   const name = game.i18n.localize('IRONSWORN.Custom Oracles')
   const rootFolder = game.tables?.directory?.folders.find(
     (x) => x.name === name
   )
   if (!rootFolder) return
 
-  function walkFolder(parent: IOracleTreeNode, folder: Folder) {
+  function walkFolder(parent: OracleTreeNode, folder: Folder) {
     // Add this folder
-    const newNode: IOracleTreeNode = {
+    const newNode: OracleTreeNode = {
       ...emptyNode(),
       displayName: folder.name || '(folder)',
     }
@@ -176,7 +173,7 @@ async function augmentWithFolderContents(node: IOracleTreeNode) {
   walkFolder(node, rootFolder)
 }
 
-function walkAndFreezeTables(node: IOracleTreeNode) {
+function walkAndFreezeTables(node: OracleTreeNode) {
   ;(node.tables as any) = Object.freeze(node.tables)
   for (const child of node.children) {
     walkAndFreezeTables(child)
@@ -184,11 +181,11 @@ function walkAndFreezeTables(node: IOracleTreeNode) {
 }
 
 export function findPathToNodeByTableId(
-  rootNode: IOracleTreeNode,
+  rootNode: OracleTreeNode,
   tableId: string
-): IOracleTreeNode[] {
-  const ret: IOracleTreeNode[] = []
-  function walk(node: IOracleTreeNode) {
+): OracleTreeNode[] {
+  const ret: OracleTreeNode[] = []
+  function walk(node: OracleTreeNode) {
     ret.push(node)
     const foundTable = node.tables.find((x) => x.id === tableId)
     if (foundTable) return true