+++ src/module/helpers/settings.ts
@@ -78,6 +78,15 @@ export class IronswornSettings {
       default: true,
     })
 
+    game.settings.register('foundry-ironsworn', 'df-is-moves-oracles', {
+      name: 'IRONSWORN.Settings.DataforgedIronswornMoves.Name',
+      hint: 'IRONSWORN.Settings.DataforgedIronswornMoves.Hint',
+      scope: 'client',
+      config: true,
+      type: Boolean,
+      default: false,
+    })
+
     game.settings.register('foundry-ironsworn', 'data-version', {
       scope: 'world',
       config: false,
@@ -106,7 +115,11 @@ export class IronswornSettings {
   }
 
   static get logCharacterChanges(): boolean {
-    return game.settings.get('foundry-ironsworn', 'log-changes') as boolean
+    return !!game.settings.get('foundry-ironsworn', 'log-changes')
+  }
+
+  static get dataforgedIronswornMoves(): boolean {
+    return !!game.settings.get('foundry-ironsworn', 'df-is-moves-oracles')
   }
 
   static async maybeSetGlobalSupply(value: number) {