+++ src/module/vue/components/character-sheet-tabs/sf-progresses.vue
@@ -1,0 +1,7 @@
+<template>
+  <active-completed-progresses />
+</template>
+
+<script setup lang="ts">
+import ActiveCompletedProgresses from '../active-completed-progresses.vue'
+</script>