+++ src/module/vue/components/conditions/condition-checkbox.vue
@@ -10,7 +10,7 @@
 </template>
 
 <script lang="ts" setup>
-import { inject, Ref } from 'vue'
+import { inject, nextTick, Ref } from 'vue'
 import { IronswornSettings } from '../../../helpers/settings'
 import { $ActorKey } from '../../provisions'
 
@@ -24,18 +24,24 @@ const props = defineProps<{
 
 async function input(ev) {
   const value = ev.currentTarget.checked
-  let numDebilitiesMarked =
-    Object.values(actor.value.data.debility).filter((x) => x).length +
-    (value ? 1 : -1)
   await $actor?.update({
     data: {
       debility: {
         [props.name]: value,
       },
+    },
+  })
+  await nextTick()
+  const numDebilitiesMarked = Object.values(actor.value.data.debility).filter(
+    (x) => x === true
+  ).length
+  await $actor?.update({
+    data: {
       momentumMax: 10 - numDebilitiesMarked,
       momentumReset: Math.max(0, 2 - numDebilitiesMarked),
     },
   })
+
   if (props.global) {
     await IronswornSettings.maybeSetGlobalCondition(props.name, value)
   }