+++ src/module/vue/components/sf-movesheetmoves.vue
@@ -62,13 +62,21 @@ h2 {
 <script setup lang="ts">
 import { flatten } from 'lodash'
 import { computed, inject, nextTick, reactive, ref, Ref } from 'vue'
-import { createStarforgedMoveTree } from '../../features/custommoves'
+import {
+  createIronswornMoveTree,
+  createStarforgedMoveTree,
+} from '../../features/custommoves'
 import { $EmitterKey } from '../provisions'
 import sfMoverow from './sf-moverow.vue'
 
+const props = defineProps<{ toolset: 'ironsworn' | 'starforged' }>()
+
 const actor = inject('actor') as Ref
 
-const tempCategories = await createStarforgedMoveTree()
+const tempCategories =
+  props.toolset === 'ironsworn'
+    ? await createIronswornMoveTree()
+    : await createStarforgedMoveTree()
 for (const category of tempCategories) {
   for (const move of category.moves) {
     move.highlighted = false